requirements.txt
dpkt==1.9.2
gpxpy==1.3.4
mysql-connector==2.1.7
netifaces==0.10.9
SQLAlchemy==1.3.3
XlsxWriter==1.1.2
matplotlib==3.0.2
geographiclib==1.49
geopy==1.18.1
numpy==1.16.2
scipy==1.2.1
